+++ scripts/aint/create_aint_model.py
@@ -9,14 +9,14 @@ from hari_client.utils import logger
 log = logger.setup_logger(__name__)
 
 
-def create_training_set(
+def create_aint_learning_data(
     hari: HARIClient,
     name: str,
     dataset_id: uuid.UUID,
     attribute_id: uuid.UUID,
     subset_id: uuid.UUID,
     user_group: str = None,
-) -> models.TrainingSet:
+) -> models.AINTLearningData:
     # construct training attributes
     if subset_id is None:
         training_attribute = models.TrainingAttribute(
@@ -33,14 +33,14 @@ def create_training_set(
             ],
         )
 
-    return hari.create_training_set(
+    return hari.create_aint_learning_data(
         name=name, training_attributes=[training_attribute], user_group=user_group
     )
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
-        description="Create and train an AI Nano Task (model). This also create the needed training sets."
+        description="Create and train an AI Nano Task (model). This also create the needed AINT learning data."
     )
 
     parser.add_argument(
@@ -85,9 +85,9 @@ if __name__ == "__main__":
     config: Config = Config(_env_file=".env")
     hari: HARIClient = HARIClient(config=config)
 
-    # Create Training Set
+    # Create AINT learning data
     # !!! only available to qm internal users !!!
-    training_set = create_training_set(
+    aint_learning_data = create_aint_learning_data(
         hari,
         args.name,
         args.dataset_id,
@@ -95,13 +95,11 @@ if __name__ == "__main__":
         args.subset_id,
         args.user_group,
     )
-    training_set_id = training_set.id
-    log.info(f"Created training set with ID: {training_set_id}")
+    aint_learning_data_id = aint_learning_data.id
+    log.info(f"Created aint learning data with ID: {aint_learning_data_id}")
 
     # Start AINT model training
-    model_id = hari.train_ml_annotation_model(
-        name=args.name, training_set_id=training_set_id
-    )
+    model_id = hari.train_ml_annotation_model(args.name, aint_learning_data_id)
 
     log.info(
         "The AINT training can take a while, please wait. "